delete ExtractorCredentialPolicy and rename ExtractorCredentialPolicyTests to CredentialPolicyGenerateTests and add some tests
